requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
pyzbar==0.1.9
Pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
aiohttp==3.9.1 